pyproject.toml
[project]
name = "embody-serial"
version = "1.0.36"
description = "Communicate with the embody device over a serial port"
authors = [{name = "Aidee Health AS", email = "hello@aidee.io"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Development Status :: 5 - Production/Stable",
]

dependencies= [
    "embody-codec>=1.0.32",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "Pygments>=2.10.0",
    "ruff>=0.11.5",
    "ty>=0.0.2",
    "pre-commit>=2.16.0",
    "pre-commit-hooks>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/aidee-health/embody-serial"
Repository = "https://github.com/aidee-health/embody-serial"
Changelog = "https://github.com/aidee-health/embody-serial/releases"

[project.scripts]
embody-serial = "embodyserial.cli:main"

[tool.setuptools.entry-points."console_scripts"]
embody-serial = "embodyserial.cli:main"

[tool.hatch.build.targets.sdist]
include = ["src/embodyserial"]

[tool.hatch.build.targets.wheel]
include = ["src/embodyserial"]

[tool.hatch.build.targets.wheel.sources]
"src/embodyserial" = "embodyserial"

[tool.pytest.ini_options]
markers = [
    "fast: Fast unit tests without EmbodySerial initialization (< 1s)",
    "timing: Tests that use intentional delays to verify timing/threading behavior",
    "lifecycle: Tests focused on initialization, connection, and shutdown behavior",
    "api: Tests of the helper API and message handling",
    "error_handling: Tests focused on exception handling and validation",
]

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["embodyserial", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 2

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.ty.environment]
python-version = "3.11"  # Matches project requires-python

[tool.ty.rules]
# Core type safety rules
unresolved-reference = "error"
invalid-assignment = "error"
invalid-argument-type = "error"
call-non-callable = "error"

# Runtime safety (Tier 1)
invalid-return-type = "error"
missing-argument = "error"
too-many-positional-arguments = "error"
unknown-argument = "error"
non-subscriptable = "error"
not-iterable = "error"

# Exception handling (Tier 1)
invalid-exception-caught = "error"
invalid-raise = "error"

# Class safety (Tier 1)
invalid-method-override = "error"
invalid-base = "error"

# Attribute access (Tier 1)
unresolved-import = "error"
unresolved-attribute = "error"

# Indexing and slicing (Tier 2)
index-out-of-bounds = "error"
zero-stepsize-in-slice = "error"

# Generic types (Tier 2)
invalid-type-arguments = "error"
invalid-generic-class = "error"

# Code quality warnings
deprecated = "warn"
redundant-cast = "warn"
unused-ignore-comment = "warn"
possibly-unresolved-reference = "warn"
ambiguous-protocol-member = "warn"

[tool.ty.terminal]
output-format = "full"             # Provides detailed error messages with full context
error-on-warning = false           # Don't fail on warnings initially

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "C",    # flake8-comprehensions
    "D",    # pydocstyle
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "N",    # pep8-naming
    "S",    # flake8-bandit
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "RUF",  # Ruff-specific rules
    "PTH",  # use pathlib
    "PL",   # Pylint
]
ignore = [
    "E203",   # Whitespace before ':'
    "E501",   # Line too long
    "D107",   # Missing docstring in __init__
    "D415",   # First line should end with a period, question mark, or exclamation point
    "D102",   # Missing docstring in public method
    "S101",   # Use of assert detected
    "D105",   # Missing docstring in magic method
    "D103",   # Missing docstring in public function
    "D101",   # Missing docstring in public class
    "B905",   # zip() without an explicit strict= parameter
    "PLR2004", # Magic value used in comparison
    "PLR0913", # Too many arguments
    "I001",   # isort: skip
    "PTH123", # pathlib: skip
    "RUF012", # mutable-class-default
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 50

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
